Remove resharding driver
