+++ lib/collection/src/collection/collection_ops.rs
@@ -380,8 +380,7 @@ impl Collection {
         }
         let shard_transfers =
             shards_holder.get_shard_transfer_info(&*self.transfer_tasks.lock().await);
-        let resharding_operations =
-            shards_holder.get_resharding_operations_info(&*self.reshard_tasks.lock().await);
+        let resharding_operations = shards_holder.get_resharding_operations_info();
 
         // sort by shard_id
         local_shards.sort_by_key(|k| k.shard_id);
